+++ build.ps1
@@ -1,4 +1,4 @@
-$version = "7.51"
+$version = "7.53"
 $templateContent = Get-Content -Path "src/template.js" -Raw
 
 $outputFile = "output/hobo-helper-v${version}.user.js"